requirements.txt
# 2025CoinTrader 통합 의존성 파일
# V1 (딥러닝) + V2 (머신러닝) 시스템을 모두 지원

# ====================================
# 핵심 데이터 처리 라이브러리
# ====================================
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# ====================================
# 머신러닝 및 딥러닝
# ====================================
# 기본 머신러닝 (V2 필수)
scikit-learn>=1.1.0
joblib>=1.1.0

# 딥러닝 (V1 필수)
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# 선택적 - 더 나은 성능을 위해
xgboost>=1.6.0

# ====================================
# 데이터 시각화
# ====================================
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# ====================================
# 데이터 I/O 및 포맷
# ====================================
# Parquet 파일 지원 (V2 필수)
pyarrow>=8.0.0
# 추가 데이터 포맷
openpyxl>=3.0.10
xlrd>=2.0.1

# ====================================
# 암호화폐 API 및 네트워킹
# ====================================
requests>=2.28.0
urllib3>=1.26.0
# 업비트 API (선택적)
pyupbit>=0.2.0

# ====================================
# 유틸리티 및 도구
# ====================================
# 환경 변수 관리
python-dotenv>=0.19.0
# 진행 표시바
tqdm>=4.64.0
# 로깅 개선
loguru>=0.6.0
# 타임존 처리
pytz>=2022.1

# ====================================
# 개발 및 테스팅 (선택적)
# ====================================
# 코드 품질
black>=22.0.0
flake8>=4.0.0
# 테스팅
pytest>=7.0.0
pytest-cov>=3.0.0

# ====================================
# 웹 기반 모니터링 (선택적)
# ====================================
flask>=2.1.0
dash>=2.5.0
streamlit>=1.10.0

# ====================================
# GPU 가속 설치 안내
# ====================================
# GPU 지원을 위해서는 다음 명령어 실행:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# 
# CUDA 버전 확인:
# python -c "import torch; print(torch.cuda.is_available(), torch.version.cuda)"

# ====================================
# 설치 후 확인 방법
# ====================================
# V1 시스템: cd V1_Based_HKM && python tools/test_setup.py
# V2 시스템: cd V2_Based_with_CHATGPT && python -c "from tools.config import Config; print('V2 설정 완료')"